Extract RestServiceTracker
The service tracker in the activation method of the service registry can
cause troubles depending on the order in which the services are started.
